Move project queries into their own hook
